+++ java-security/src/main/java/com/sap/cloud/security/token/validation/validators/JwtAudienceValidator.java
@@ -42,25 +42,15 @@ public class JwtAudienceValidator implements Validator<Token> {
 
 	@Override
 	public ValidationResult validate(Token token) {
-		Set<String> allowedAudiences = getAllowedAudiences(token);
+		Set<String> allowedAudiences = extractAudiencesFromToken(token);
 
-		if(validateSameClientId(token) || validateDefault(allowedAudiences)
+		if (validateDefault(allowedAudiences)
 				|| validateAudienceOfXsuaaBrokerClone(allowedAudiences)) {
 			return ValidationResults.createValid();
 		}
 		return ValidationResults.createInvalid(
 				"Jwt token with audience {} is not issued for these clientIds: {}.",
-				allowedAudiences, clientIds);
-	}
-
-	private boolean validateSameClientId(Token token) {
-		if(Service.XSUAA.equals(token.getService()) && token.hasClaim(TokenClaims.XSUAA.CLIENT_ID)) {
-			String clientId = token.getClaimAsString(TokenClaims.XSUAA.CLIENT_ID);
-			if(clientIds.contains(clientId)) {
-				return true;
-			}
-		}
-		return false;
+				token.getAudiences(), clientIds);
 	}
 
 	private boolean validateDefault(Set<String> allowedAudiences) {
@@ -76,7 +66,7 @@ public class JwtAudienceValidator implements Validator<Token> {
 		for (String configuredClientId : clientIds) {
 			if (configuredClientId.contains("!b")) {
 				for (String audience : allowedAudiences) {
-					if (audience.contains("|") && audience.endsWith("|" + configuredClientId)) {
+					if (audience.endsWith("|" + configuredClientId)) {
 						return true;
 					}
 				}
@@ -91,7 +81,7 @@ public class JwtAudienceValidator implements Validator<Token> {
 	 * @param token
 	 * @return (empty) list of audiences
 	 */
-	static Set<String> getAllowedAudiences(Token token) {
+	static Set<String> extractAudiencesFromToken(Token token) {
 		Set<String> audiences = new LinkedHashSet<>();
 
 		for (String audience : token.getAudiences()) {
@@ -106,20 +96,28 @@ public class JwtAudienceValidator implements Validator<Token> {
 				audiences.add(audience);
 			}
 		}
-		// extract audience (app-id) from scopes
-		if (audiences.isEmpty() && Service.XSUAA.equals(token.getService())) {
-			for (String scope : token.getClaimAsStringList(TokenClaims.XSUAA.SCOPES)) {
-				if (scope.contains(".")) {
-					audiences.add(extractAppId(scope));
+
+		if (Service.XSUAA.equals(token.getService())) {
+			if (token.hasClaim(TokenClaims.XSUAA.CLIENT_ID)) {
+				audiences.add(token.getClaimAsString(TokenClaims.XSUAA.CLIENT_ID));
+			}
+			// extract audience (app-id) from scopes
+			if (token.getAudiences().isEmpty()) {
+				for (String scope : token.getClaimAsStringList(TokenClaims.XSUAA.SCOPES)) {
+					if (scope.contains(".")) {
+						audiences.add(extractAppId(scope));
+					}
 				}
 			}
 		}
+		logger.info("The audiences that are derived from the token: {}.", audiences);
 		return audiences;
 	}
 
 	/**
-	 * In case of audiences, the namespaces are trimmed.
-	 * In case of scopes, the namespaces and the scope names are trimmed.
+	 * In case of audiences, the namespaces are trimmed. In case of scopes, the
+	 * namespaces and the scope names are trimmed.
+	 * 
 	 * @param scopeOrAudience
 	 * @return
 	 */